Commit initial code
Adds a (completely functional) first version of this module, made up of
an index.js and two files it passes work onto: the loader and the
evaluator.
The main export has to be called with a file path, properties to
substitute in said file, and an options object. It can sanitize HTML
values and normalize line endings, if specified in the options. Once a
template is read, it is cached for faster access in the future (can be
turned off).